+++ src/php/Gdbots/Schemas/Enrichments/Mixin/Utm/UtmV1.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Gdbots\Schemas\Enrichments\Mixin\Utm;
+
+interface UtmV1 extends Utm
+{
+}