+++ src/php/Gdbots/Schemas/Enrichments/Mixin/Utm/UtmV1Trait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Gdbots\Schemas\Enrichments\Mixin\Utm;
+
+use Gdbots\Pbj\Schema;
+
+/**
+ * @method static Schema schema
+ * @method mixed get($fieldName, $default = null)
+ */
+trait UtmV1Trait
+{
+}