+++ src/php/Gdbots/Schemas/Pbjx/Event/EventExecutionFailedV1.php
@@ -10,20 +10,25 @@ use Gdbots\Schemas\Pbjx\Enum\Code;
 use Gdbots\Schemas\Pbjx\Mixin\Event\EventV1;
 use Gdbots\Schemas\Pbjx\Mixin\Event\EventV1Mixin;
 use Gdbots\Schemas\Pbjx\Mixin\Event\EventV1Trait;
+use Gdbots\Schemas\Pbjx\Mixin\Indexed\IndexedV1;
+use Gdbots\Schemas\Pbjx\Mixin\Indexed\IndexedV1Mixin;
+use Gdbots\Schemas\Pbjx\Mixin\Indexed\IndexedV1Trait;
 
 final class EventExecutionFailedV1 extends AbstractMessage implements
     EventExecutionFailed,
-    EventV1
+    EventV1,
+    IndexedV1
   
 {
     use EventV1Trait;
+    use IndexedV1Trait;
 
     /**
      * @return Schema
      */
     protected static function defineSchema()
     {
-        return new Schema('pbj:gdbots:pbjx:event:event-execution-failed:1-0-0', __CLASS__,
+        return new Schema('pbj:gdbots:pbjx:event:event-execution-failed:1-0-1', __CLASS__,
             [
                 Fb::create('event', T\MessageType::create())
                     ->className('Gdbots\Schemas\Pbjx\Mixin\Event\Event')
@@ -38,7 +43,8 @@ final class EventExecutionFailedV1 extends AbstractMessage implements
                     ->build()
             ],
             [
-                EventV1Mixin::create()
+                EventV1Mixin::create(), 
+                IndexedV1Mixin::create()
             ]
         );
     }